Removed EXPECT_TRUE from subroutine setDimensionsAndCheckBounds()
gtest has an issue with propagating failures as described in
https://github.com/google/googletest/blob/master/googletest/docs/AdvancedGuide.md#propagating-fatal-failures
However, the EXPECT_TRUE in the subroutine is not necessary since the subroutine
returns true iff that check passes, and invocations of the subroutine have
assertions on the output. Therefore the conditions under which the tests pass
are unchanged by removing this line.