Remove unnecessary dependencies for ${LIB_STATIC_NUPICCORE_COMBINED}
The only build target that is an input to this build step is
${LIB_STATIC_NUPICCORE}, so that is all it should depend on. All the
other inputs are determined at configure time and reside in
`nupic.core/external`.

I noticed this when touching
`nupic.core/src/test/integration/ConnectionsPerformanceTest.cpp`
which is needed for the ${EXECUTABLE_CONNECTIONSPERFORMANCETEST} target.
This would cause ${LIB_STATIC_NUPICCORE_COMBINED} to be rebuilt as well,
which would also trigger rebuilding the SWIG bindings (which takes a
while).