a fresh look at ExcitedColumns: it's just a group_by
Adds a group_by. Gets rid of ExcitedColumns and DepolarizedCells.

This is doing fundamentally the same thing, but it's now described
using a reusable concept: the "group by".

The code is more verbose, but there's less conceptual overhead that's
specific to the Temporal Memory. And it encourages better practices --
the previous code would encourage you to reuse the ExcitedColumns and
DepolarizedCells even when you only cared about some of the sequences
that it traversed (e.g. just the active segments).

According to temporal_memory_performance_benchmark.py, this change is
performance-neutral.