Stop putting tie-break info in the overlaps
Rather than making a copy of the overlaps and using the same logic,
move away from doing the tie-break logic with math and an
"arbitration" constant. Instead, handle the tie-break case explicitly,
checking if the neighboring column is selected in the event of a tie.

This change makes the code less error-prone because:

1. We're not corrupting the overlaps (which are exposed via
   getBoostedOverlaps)
2. We're not having to use a heuristic to choose an arbitration
   constant.