Pybind11 SpatialPooler fix for printing durring constructor.
https://pybind11.readthedocs.io/en/stable/advanced/pycpp/utilities.html#capturing-standard-output-from-ostream

Often, a library will use the streams std::cout and std::cerr to
print, but this does not play well with Python’s standard sys.stdout
and sys.stderr redirection. Replacing a library’s printing with
py::print may not be feasible. This can be fixed using a guard
around the library function that redirects output to the
corresponding Python streams ...