Fixed particle swarming optimization
Added a mutex lock to the ParticleData so that it does not get corrupted
by being evaluated twice at the same time.  Because the optimizer always
runs from the main thread, it uses a simple flag (instead of a
thread safe mutex).

Also fixed the python MNIST example.  Now it works with the optimization
framework, and also returns a good value for CI to use.