Connections: createSynapse() checks for duplicit presynaptic synapses
The logic was moved from TM.growSynapses_() to
Connections.createSynapse() as it's generally applicable and correct.

We want to avoid creating synapses to already existing presynaptic cells
(on the same segment) as that breaks the binary nature of synapses.

Advantage of this is simplified implementation and much cleaner code in
TM.