Remove the -Werror flag.
The compiler flag "-Werror" converts all warnings into errors.
This was good back when we had CI & prebuilt binaries,
but now that the best advice is for users to build from source
this is causing problems. Deprecation warnings in particular
are making the library un-installable, despite the fact that the
deprecated functions are still present.